+++ lib/Events/V1BillingMeterErrorReportTriggeredEventNotification.php
@@ -7,7 +7,7 @@ namespace Stripe\Events;
 /**
  * @property \Stripe\RelatedObject $related_object Object containing the reference to API resource relevant to the event
  */
-class V1BillingMeterErrorReportTriggeredEventNotification extends \Stripe\V2\EventNotification
+class V1BillingMeterErrorReportTriggeredEventNotification extends \Stripe\V2\Core\EventNotification
 {
     const LOOKUP_TYPE = 'v1.billing.meter.error_report_triggered';
     public $related_object;
