+++ lib/Events/V1BillingMeterNoMeterFoundEvent.php
@@ -7,7 +7,7 @@ namespace Stripe\Events;
 /**
  * @property \Stripe\EventData\V1BillingMeterNoMeterFoundEventData $data data associated with the event
  */
-class V1BillingMeterNoMeterFoundEvent extends \Stripe\V2\Event
+class V1BillingMeterNoMeterFoundEvent extends \Stripe\V2\Core\Event
 {
     const LOOKUP_TYPE = 'v1.billing.meter.no_meter_found';
