+++ lib/Events/V1BillingMeterNoMeterFoundEventNotification.php
@@ -4,7 +4,7 @@
 
 namespace Stripe\Events;
 
-class V1BillingMeterNoMeterFoundEventNotification extends \Stripe\V2\EventNotification
+class V1BillingMeterNoMeterFoundEventNotification extends \Stripe\V2\Core\EventNotification
 {
     const LOOKUP_TYPE = 'v1.billing.meter.no_meter_found';
