+++ lib/Service/V2/Core/EventDestinationService.php
@@ -17,7 +17,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array{include?: string[], limit?: int} $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\Collection<\Stripe\V2\EventDestination>
+     * @return \Stripe\V2\Collection<\Stripe\V2\Core\EventDestination>
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -32,7 +32,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array{description?: string, enabled_events: string[], event_payload: string, events_from?: string[], include?: string[], metadata?: array<string, string>, name: string, snapshot_api_version?: string, type: string, amazon_eventbridge?: array{aws_account_id: string, aws_region: string}, webhook_endpoint?: array{url: string}} $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\EventDestination
+     * @return \Stripe\V2\Core\EventDestination
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -64,7 +64,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\EventDestination
+     * @return \Stripe\V2\Core\EventDestination
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -80,7 +80,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\EventDestination
+     * @return \Stripe\V2\Core\EventDestination
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -96,7 +96,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\Event
+     * @return \Stripe\V2\Core\Event
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -112,7 +112,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array{include?: string[]} $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\EventDestination
+     * @return \Stripe\V2\Core\EventDestination
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
@@ -128,7 +128,7 @@ class EventDestinationService extends \Stripe\Service\AbstractService
      * @param null|array{description?: string, enabled_events?: string[], include?: string[], metadata?: array<string, null|string>, name?: string, webhook_endpoint?: array{url: string}} $params
      * @param null|RequestOptionsArray|\Stripe\Util\RequestOptions $opts
      *
-     * @return \Stripe\V2\EventDestination
+     * @return \Stripe\V2\Core\EventDestination
      *
      * @throws \Stripe\Exception\ApiErrorException if the request fails
      */
