+++ lib/Util/Util.php
@@ -69,7 +69,7 @@ abstract class Util
                     if (\array_key_exists('type', $resp) && \array_key_exists($resp['type'], $eventTypes)) {
                         $class = $eventTypes[$resp['type']];
                     } else {
-                        $class = \Stripe\V2\Event::class;
+                        $class = \Stripe\V2\Core\Event::class;
                     }
                 }
             } elseif (\array_key_exists('data', $resp) && \array_key_exists('next_page_url', $resp)) {
