+++ lib/V2/Core/Event.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Stripe\V2;
+namespace Stripe\V2\Core;
 
 /**
  * Base class for V2 events.
  *
  * This is concrete for use in our generated tests. Events returned from the \Stripe\V2\Core\EventService
- * will be a subtype of \Stripe\V2\Event.
+ * will be a subtype of \Stripe\V2\Core\Event.
  *
  * @property string $id Unique identifier for the event.
  * @property string $object String representing the object's type. Objects of the same type share the same value of the object field.
