+++ lib/V2/Core/EventNotification.php
@@ -1,9 +1,10 @@
 <?php
 
-namespace Stripe\V2;
+namespace Stripe\V2\Core;
 
 use Stripe\Reason;
 use Stripe\RelatedObject;
+use Stripe\Events\UnknownEventNotification;
 use Stripe\Util\EventNotificationTypes;
 
 /**
