+++ tests/Stripe/EventTest.php
@@ -4,14 +4,14 @@ namespace Stripe;
 
 use Stripe\Events\V1BillingMeterErrorReportTriggeredEventNotification;
 use Stripe\Events\V1BillingMeterNoMeterFoundEventNotification;
-use Stripe\V2\EventNotification;
-use Stripe\V2\UnknownEventNotification;
+use Stripe\V2\Core\EventNotification;
+use Stripe\Events\UnknownEventNotification;
 
 /**
  * @internal
  *
  * @covers \Stripe\Event
- * @covers \Stripe\V2\Event
+ * @covers \Stripe\V2\Core\Event
  */
 final class EventTest extends TestCase
 {
