+++ tests/Stripe/GeneratedExamplesTest.php
@@ -607,7 +607,7 @@ final class GeneratedExamplesTest extends TestCase
             BaseStripeClient::DEFAULT_API_BASE
         );
         $result = $this->v2Client->v2->core->events->retrieve('ll_123', []);
-        self::assertInstanceOf(V2\Event::class, $result);
+        self::assertInstanceOf(V2\Core\Event::class, $result);
     }
 
     public function testCountrySpecsGet()
@@ -4955,7 +4955,7 @@ final class GeneratedExamplesTest extends TestCase
             'object_id' => 'object_id',
         ]);
         self::assertInstanceOf(V2\Collection::class, $result);
-        self::assertInstanceOf(V2\Event::class, $result->data[0]);
+        self::assertInstanceOf(V2\Core\Event::class, $result->data[0]);
     }
 
     public function testV2CoreEventGet2()
@@ -4977,7 +4977,7 @@ final class GeneratedExamplesTest extends TestCase
             BaseStripeClient::DEFAULT_API_BASE
         );
         $result = $this->v2Client->v2->core->events->retrieve('id_123', []);
-        self::assertInstanceOf(V2\Event::class, $result);
+        self::assertInstanceOf(V2\Core\Event::class, $result);
     }
 
     public function testV2CoreEventDestinationGet()
@@ -5012,7 +5012,7 @@ final class GeneratedExamplesTest extends TestCase
         );
         $result = $this->v2Client->v2->core->eventDestinations->all([]);
         self::assertInstanceOf(V2\Collection::class, $result);
-        self::assertInstanceOf(V2\EventDestination::class, $result->data[0]);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result->data[0]);
     }
 
     public function testV2CoreEventDestinationPost()
@@ -5050,7 +5050,7 @@ final class GeneratedExamplesTest extends TestCase
             'name' => 'name',
             'type' => 'amazon_eventbridge',
         ]);
-        self::assertInstanceOf(V2\EventDestination::class, $result);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result);
     }
 
     public function testV2CoreEventDestinationDelete()
@@ -5103,7 +5103,7 @@ final class GeneratedExamplesTest extends TestCase
             'id_123',
             []
         );
-        self::assertInstanceOf(V2\EventDestination::class, $result);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result);
     }
 
     public function testV2CoreEventDestinationPost2()
@@ -5134,7 +5134,7 @@ final class GeneratedExamplesTest extends TestCase
             'id_123',
             []
         );
-        self::assertInstanceOf(V2\EventDestination::class, $result);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result);
     }
 
     public function testV2CoreEventDestinationPost3()
@@ -5165,7 +5165,7 @@ final class GeneratedExamplesTest extends TestCase
             'id_123',
             []
         );
-        self::assertInstanceOf(V2\EventDestination::class, $result);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result);
     }
 
     public function testV2CoreEventDestinationPost4()
@@ -5196,7 +5196,7 @@ final class GeneratedExamplesTest extends TestCase
             'id_123',
             []
         );
-        self::assertInstanceOf(V2\EventDestination::class, $result);
+        self::assertInstanceOf(V2\Core\EventDestination::class, $result);
     }
 
     public function testV2CoreEventDestinationPost5()
@@ -5221,7 +5221,7 @@ final class GeneratedExamplesTest extends TestCase
             'id_123',
             []
         );
-        self::assertInstanceOf(V2\Event::class, $result);
+        self::assertInstanceOf(V2\Core\Event::class, $result);
     }
 
     public function testTemporarySessionExpiredError()
